# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agent ID — Identity, trust, routing, and commerce layer for AI agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List user's agents
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Agent"
                required: [agents]
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentInput"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
...
  /v1/programmatic/agents/register:
    post:
      operationId: programmaticRegister
      tags: [programmatic]
      summary: Programmatic agent registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgrammaticRegisterInput"
      responses:
        "201":
          description: Registration initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgrammaticRegisterResponse"
  /v1/programmatic/agents/verify:
    post:
      operationId: programmaticVerify
      tags: [programmatic]
      summary: Verify agent key control
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgrammaticVerifyInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgrammaticVerifyResponse"

  # --- Handles ---
...
  /v1/agents/{agentId}/verify/initiate:
    post:
      operationId: initiateVerification
      tags: [verification]
      summary: Initiate agent verification
      parameters:
        - $ref: "#/components/parameters/AgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
              required: [method]
      responses:
        "200":
          description: Verification challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationChallenge"
...
  /v1/billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutInput"
      responses:
        "200":
          description: Checkout session
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkoutUrl:
                    type: string
                required: [checkoutUrl]
  /v1/billing/agents/{agentId}/activate:
    post:
      operationId: activateAgent
      tags: [billing]
      summary: Activate agent on monthly plan
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Activation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSubscription"
  /v1/billing/agents/{agentId}/deactivate:
    post:
      operationId: deactivateAgent
      tags: [billing]
      summary: Deactivate agent
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Deactivation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSubscription"
  /v1/billing/agents/{agentId}/status:
    get:
      operationId: getAgentBillingStatus
      tags: [billing]
      summary: Get agent billing status
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSubscription"

  # --- Payments ---
...
components:
  parameters:
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        handle:
          type: string
        displayName:
          type: string
        description:
          type: string
        avatarSeed:
          type: string
        avatarUrl:
          type: string
        status:
          type: string
        isPublic:
          type: boolean
        endpointUrl:
          type: string
        capabilities:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        protocols:
          type: array
          items:
            type: string
        authMethods:
          type: array
          items:
            type: string
        paymentMethods:
          type: array
          items:
            type: string
        trustScore:
          type: integer
        trustBreakdown:
          type: object
        trustTier:
          type: string
        verificationStatus:
          type: string
        verificationMethod:
          type: string
        verifiedAt:
          type: string
          format: date-time
        tasksReceived:
          type: integer
        tasksCompleted:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, handle, displayName, status, trustScore, trustTier, verificationStatus, createdAt]
    CreateAgentInput:
      type: object
      properties:
        handle:
          type: string
        displayName:
          type: string
        description:
          type: string
        endpointUrl:
          type: string
        capabilities:
          type: array
          items:
            type: string
        protocols:
          type: array
          items:
            type: string
      required: [handle, displayName]
...
    ProgrammaticRegisterInput:
      type: object
      properties:
        handle:
          type: string
        displayName:
          type: string
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
        endpointUrl:
          type: string
        publicKeys:
          type: array
          items:
            type: object
            properties:
              kid:
                type: string
              kty:
                type: string
              publicKey:
                type: string
              use:
                type: string
            required: [kid, kty, publicKey]
        authMethods:
          type: array
          items:
            type: string
        protocols:
          type: array
          items:
            type: string
        paymentMethods:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
      required: [handle, displayName, publicKeys]
    ProgrammaticRegisterResponse:
      type: object
      properties:
        agentId:
          type: string
          format: uuid
        handle:
          type: string
        domain:
          type: string
        verificationChallenge:
          type: string
        challengeExpiresAt:
          type: string
          format: date-time
        registrationStatus:
          type: string
        profileUrl:
          type: string
      required: [agentId, handle, verificationChallenge, challengeExpiresAt, registrationStatus]
    ProgrammaticVerifyInput:
      type: object
      properties:
        agentId:
          type: string
          format: uuid
        challenge:
          type: string
        kid:
          type: string
        signature:
          type: string
        timestamp:
          type: string
          format: date-time
        method:
          type: string
      required: [agentId, challenge, kid, signature]
    ProgrammaticVerifyResponse:
      type: object
      properties:
        status:
          type: string
        agentId:
          type: string
          format: uuid
        handle:
          type: string
        domain:
          type: string
        profileUrl:
          type: string
        trustScore:
          type: integer
        trustTier:
          type: string
        domainStatus:
          type: string
      required: [status, agentId, handle, trustScore, trustTier]
...
    VerificationChallenge:
      type: object
      properties:
        challengeId:
          type: string
          format: uuid
        challenge:
          type: string
        method:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [challengeId, challenge, method, expiresAt]
...
    AgentSubscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agentId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        plan:
          type: string
        status:
          type: string
        provider:
          type: string
        billingInterval:
          type: string
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, agentId, userId, plan, status, createdAt]
    CreateCheckoutInput:
      type: object
      properties:
        plan:
          type: string
        billingInterval:
          type: string
        agentId:
          type: string
          format: uuid
      required: [plan]
...